Fix testing with `celo` build tag
Several tests caused conflicts with the
`go-ethereum/crypto` library. Solved the problem
using build tags and moving some code.
