+++ pkg/ecdsa/tss/tss_ethereum_test.go
@@ -1,0 +1,11 @@
+//+build !celo
+
+package tss
+
+import (
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/crypto/secp256k1"
+)
+
+var S256 = secp256k1.S256
+var SigToPub = crypto.SigToPub
